requirements.txt
pytest
selenium
pytest-faker
requests
pytest-html
pytest-html-reporter
allure-pytest
pytest-rerunfailures